+++ esphome/components/nextion/defines.py
@@ -1,1 +1,3 @@
-CONF_VARIABLE_ID = "variable_id"
+CONF_VARIABLE_NAME = "variable_name"
+CONF_BUTTON_ID = "button_id"
+CONF_COMPONENT_NAME = "component_name"
