+++ esphome/components/nextion/nextion.cpp
@@ -21,8 +21,15 @@ void Nextion::setup() {
   this->send_command_printf("bkcmd=3");
   this->set_backlight_brightness(static_cast<uint8_t>(brightness_ * 100));
   this->goto_page("0");
+
   for (auto *sensortype : this->sensortype_) {
-    sensortype->restore_from_nextion();
+    sensortype->nextion_setup();
+  }
+  for (auto *switchtype : this->switchtype_) {
+    switchtype->nextion_setup();
+  }
+  for (auto *textsensortype : this->textsensortype_) {
+    textsensortype->nextion_setup();
   }
 }
 
@@ -269,39 +276,125 @@ bool Nextion::read_until_ack_() {
         ESP_LOGD(TAG, "Got touch at x=%u y=%u type=%s", x, y, touch_event ? "PRESS" : "RELEASE");
         break;
       }
-      case 0x66:    // sendme page id
-      case 0x70:    // string variable data return
-      case 0x71:    // numeric variable data return
-      case 0x86:    // device automatically enters into sleep mode
-      case 0x87:    // device automatically wakes up
-      case 0x88:    // system successful start up
-      case 0x89:    // start SD card upgrade
-      case 0x90: {  // Response number made to read state from Nextion
-        if (data_length != 3) {
+      case 0x66:  // sendme page id
+      case 0x70:  // string variable data return
+      case 0x71:  // numeric variable data return
+      case 0x86:  // device automatically enters into sleep mode
+      case 0x87:  // device automatically wakes up
+      case 0x88:  // system successful start up
+      case 0x89:  // start SD card upgrade
+      // Data from nextion is
+      // 0x90 - Start
+      // variable length of 0x70 return formatted data (bytes) that contain the variable name: prints "temp1",0
+      // 00 - NULL
+      // 00/01 - Single byte for on/off
+      // FF FF FF - End
+      case 0x90: {  // Switched component
+        // if (data_length != 3) {
+        //   invalid_data_length = true;
+        //   break;
+        // }
+        char variable_name[64];
+        uint8_t variable_name_end = 0;
+
+        bool value;
+        uint8_t index = 0;
+
+        // Get variable name
+        for (index = 0; index < data_length; ++index) {
+          variable_name[index] = data[index];
+          if (data[index] == 0) {  // First Null
+            variable_name_end = index;
+            break;
+          }
+        }
+        if (variable_name_end == 0) {
           invalid_data_length = true;
           break;
         }
-        uint8_t page_id = data[0];
-        uint8_t component_id = data[1];
-        uint8_t touch_event = data[2];  // 0 -> release, 1 -> press
-        ESP_LOGD(TAG, "Got special state 0x90 page=%u component=%u type=%s", page_id, component_id,
-                 touch_event ? "PRESS" : "RELEASE");
-        for (auto *touch : this->switchtype_) {
-          touch->process(page_id, component_id, touch_event);
+        ++index;
+
+        ESP_LOGD(TAG, "Got special state 0x90 variable_name=%s value=%d", variable_name, data[index]);
+        for (auto *switchtype : this->switchtype_) {
+          switchtype->process_bool(&variable_name[0], data[index] == 0 ? false : true);
         }
         break;
       }
-      case 0x91: {
-        if (data_length != 3) {
+      // Data from nextion is
+      // 0x91 - Start
+      // variable length of 0x70 return formatted data (bytes) that contain the variable name: prints "temp1",0
+      // 00 - NULL
+      // variable length of 0x71 return data: prints temp1.val,0
+      // FF FF FF - End
+      case 0x91: {  // variable/integer component
+        ESP_LOGD("nextion loop", "data_length %d", data_length);
+        char variable_name[64];
+        uint8_t variable_name_end = 0;
+
+        uint32_t value = 0;
+        uint8_t index = 0;
+
+        // Get variable name
+        for (index = 0; index < data_length; ++index) {
+          variable_name[index] = data[index];
+          if (data[index] == 0) {  // First Null
+            variable_name_end = index;
+            break;
+          }
+        }
+        if (variable_name_end == 0) {
           invalid_data_length = true;
           break;
         }
-        uint8_t page_id = data[0];
-        uint8_t component_id = data[1];
-        uint8_t touch_event = data[2];
-        ESP_LOGD(TAG, "Got special state 0x91 page=%u component=%u type=%u", page_id, component_id, touch_event);
+        // Get variable data
+        uint8_t num_byte_index = 0;
+        for (int i = index + 1; i < data_length; ++i) {
+          uint32_t factor = pow(2, (num_byte_index++ * 8));
+          uint32_t to_add = data[i] * factor;
+          value += to_add;
+        }
+        ESP_LOGD(TAG, "Got special state 0x91 variable_name=%s value=%d", variable_name, value);
         for (auto *sensor : this->sensortype_) {
-          sensor->process_sensor(page_id, component_id, touch_event);
+          sensor->process_sensor(&variable_name[0], value);
+        }
+        break;
+      }
+      // Data from nextion is
+      // 0x92 - Start
+      // variable length of 0x70 return formatted data (bytes) that contain the variable name: prints "temp1",0
+      // 00 - NULL
+      // variable length of 0x70 return formatted data (bytes) that contain the text prints temp1.txt,0
+      // 00 - NULL
+      // FF FF FF - End
+      case 0x92: {  // Switched component
+        char variable_name[64];
+        char text_value[128];
+        uint8_t variable_name_end = 0;
+        uint8_t index = 0;
+
+        // Get variable name
+        for (index = 0; index < data_length; ++index) {
+          variable_name[index] = data[index];
+          if (data[index] == 0) {  // First Null
+            variable_name_end = index;
+            break;
+          }
+        }
+        if (variable_name_end == 0) {
+          invalid_data_length = true;
+          break;
+        }
+        ++index;
+
+        for (int i = index + 1; i < data_length; ++i) {
+          text_value[index] = data[i];
+          if (data[i] == 0) {  // Second Null
+            break;
+          }
+        }
+        ESP_LOGD(TAG, "Got special state 0x92 variable_name=%s value=%d", variable_name, data[index]);
+        for (auto *textsensortype : this->textsensortype_) {
+          textsensortype->process_text(&variable_name[0], &text_value[0]);
         }
         break;
       }
@@ -313,7 +406,10 @@ bool Nextion::read_until_ack_() {
         break;
     }
     if (invalid_data_length) {
-      ESP_LOGW(TAG, "Invalid data length from nextion!");
+      ESP_LOGW(TAG, "Invalid data length (%d) from nextion!", data_length);
+      for (int index = 0; index < data_length; ++index) {
+        ESP_LOGD("nextion loop invalid_data_length", "data[%d] %d", index, data[index]);
+      }
     }
   }
 
@@ -334,35 +430,6 @@ void Nextion::set_nextion_rtc_time(time::ESPTime time) {
   this->send_command_printf("rtc5=%u", time.second);
 }
 #endif
-bool Nextion::upload_from_stream_(Stream &my_file, int content_length) {
-#if defined ESP8266
-  yield();
-#endif
-
-  // create buffer for read
-  uint8_t buff[2048] = {0};
-  // read all data from server
-  while (content_length > 0) {
-    // get available data size
-    size_t size = my_file.available();
-    if (size) {
-      int c = my_file.readBytes(buff, ((size > sizeof(buff)) ? sizeof(buff) : size));
-      if (this->debug_print_)
-        ESP_LOGD(TAG, "upload_from_stream_ sending %d bytes : total %d", c, this->total_);
-      // Write the buffered bytes to the nextion. If this fails, return false.
-      if (!this->upload_from_buffer_(buff, c)) {
-        return false;
-      }
-
-      if (content_length > 0) {
-        content_length -= c;
-      }
-    }
-    delay(1);
-  }
-
-  return true;
-}
 
 //  0x70 0x61 0x62 0x31 0x32 0x33 0xFF 0xFF 0xFF
 //  Returned when using get command for a string.
@@ -370,7 +437,7 @@ bool Nextion::upload_from_stream_(Stream &my_file, int content_length) {
 //  data: ab123
 bool Nextion::gets(const char *component_id, char *string_buffer) {
   char command[64];
-  sprintf(command, "get %s.txt", component_id);
+  sprintf(command, "get %s", component_id);
   String response = "";
   this->send_command_no_ack(command);
   this->recv_ret_string_(response);
@@ -396,311 +463,37 @@ uint32_t Nextion::getn(const char *component_id) {
   String response = "";
   uint32_t return_value = 0;
 
-  sprintf(command, "get %s.val", component_id);
+  sprintf(command, "get %s", component_id);
   this->send_command_no_ack(command);
 
   this->recv_ret_string_(response);
   if (response[0] == 0x71) {
     response.remove(0, 1);
+
     for (int i = 0; i < response.length(); ++i) {
       uint32_t factor = pow(2, (i * 8));
       uint32_t to_add = response[i] * factor;
       return_value += to_add;
     }
-    ESP_LOGD(TAG, "Received gets response (%d) for component id %s", return_value, component_id);
+
+    ESP_LOGD(TAG, "Received getn response (%d) for component id %s", return_value, component_id);
   } else {
-    ESP_LOGD(TAG, "Received unknown gets response  \"%s\" for component id %s", response.c_str(), component_id);
+    ESP_LOGD(TAG, "Received unknown getn response \"%s\" for component id %s", response.c_str(), component_id);
   }
   return return_value;
 }
 
-bool Nextion::upload_from_buffer_(const uint8_t *file_buf, size_t buf_size) {
-#if defined ESP8266
-  yield();
-#endif
-
-  uint8_t c;
-  uint8_t timeout = 0;
-  String string = String("");
-
-  for (uint16_t i = 0; i < buf_size; i++) {
-    // Users must split the .tft file contents into 4096 byte sized packets with
-    // the final partial packet size equal to the last remaining bytes (<4096
-    // bytes).
-    if (this->sent_packets_ == 4096) {
-      // wait for the Nextion to return its 0x05 byte confirming reception and
-      // readiness to receive the next packets
-      this->recv_ret_string_(string, 500, true);
-      if (string.indexOf(0x05) != -1) {
-        // reset sent packets counter
-        this->sent_packets_ = 0;
-
-        // reset receive String
-        string = "";
-      } else {
-        if (timeout >= 8) {
-          ESP_LOGD(TAG, "serial connection lost");
-          return false;
-        }
-
-        timeout++;
-      }
-
-      // delay current byte
-      i--;
-
-    } else {
-      // read buffer
-      c = file_buf[i];
-
-      // write byte to nextion over serial
-      this->write_byte(c);
-
-      // update sent packets counter
-      this->sent_packets_++;
-      this->total_++;
-      delayMicroseconds(100);
-    }
+uint32_t Nextion::nextion_71_to_uint32(String data) {
+  uint32_t return_value = 0;
+  for (int i = 0; i < data.length(); ++i) {
+    uint32_t factor = pow(2, (i * 8));
+    uint32_t to_add = data[i] * factor;
+    return_value += to_add;
   }
-
-  return true;
-}
-
-bool Nextion::upload_by_chunks_(int content_length, int chunk_size) {
-  if (this->debug_print_)
-    ESP_LOGD(TAG, "upload_by_chunks_: contentLength %d , chunk_size: %d", content_length, chunk_size);
-
-  for (int range_start = 0; range_start < content_length; range_start += chunk_size) {
-    int range_end = range_start + chunk_size - 1;
-    if (range_end > content_length)
-      range_end = content_length;
-
-    HTTPClient http;
-    if (!http.begin(this->tft_url_.c_str())) {
-      ESP_LOGD(TAG, "upload_by_chunks_: connection failed");
-      return false;
-    }
-    char range_header[64];
-    sprintf(range_header, "bytes=%d-%d", range_start, range_end);
-
-    http.addHeader("Range", range_header);
-
-    if (this->debug_print_)
-      ESP_LOGD(TAG, "upload_by_chunks_ Requesting range: %s", range_header);
-
-    // http.setReuse(true);
-    int tries = 1;
-    int code = http.GET();
-    while (code != 200 && code != 206 && tries <= 5) {
-      ESP_LOGD(TAG, "upload_by_chunks_ retrying (%d/5)", tries);
-      for (int i = 0; i < 12; ++i)  // Need a decent delay and since we will be rebooting this shouldnt be an issue.
-        delay(40);
-
-      code = http.GET();
-      ++tries;
-    }
-    if (code == 200 || code == 206) {
-      // Upload the received byte Stream to the nextion
-      bool result = this->upload_from_stream_(*http.getStreamPtr(), range_end - range_start);
-      if (result) {
-        if (this->debug_print_)
-          ESP_LOGD(TAG, "Succesfully sent chunk to Nextion");
-      } else {
-        ESP_LOGD(TAG, "upload_by_chunks_: Error updating Nextion");
-        http.end();
-        return false;
-      }
-    } else {
-      http.end();
-      return false;
-    }
-    http.end();  // End this HTTP call because we read all the data
-  }
-
-  if (content_length % 4096 != 0) {  // If not in 4096 chunks wait for the last bits to confirm
-    String string = String("");
-    uint8_t timeout = 0;
-    this->recv_ret_string_(string, 500, true);
-    if (string.indexOf(0x05) != -1) {
-      // reset sent packets counter
-      this->sent_packets_ = 0;
-
-      // reset receive String
-      string = "";
-    } else {
-      if (timeout >= 8) {
-        ESP_LOGD(TAG, "serial connection lost");
-        return false;
-      }
-
-      timeout++;
-    }
-  }
-
-  return true;
-}
-
-uint16_t Nextion::recv_ret_string_(String &response, uint32_t timeout, bool recv_flag) {
-#if defined ESP8266
-  yield();
-#endif
-
-  uint16_t ret = 0;
-  uint8_t c = 0;
-  uint8_t nr_of_FF_bytes = 0;
-  long start;
-  bool exit_flag = false;
-  bool ff_flag = false;
-  if (timeout != 500)
-    ESP_LOGD(TAG, "timeout serial read: %d", timeout);
-
-  start = millis();
-
-  while (millis() - start <= timeout) {
-    while (this->available()) {
-      this->read_byte(&c);
-      if (c == 0) {
-        continue;
-      }
-
-      if (c == 0xFF)
-        nr_of_FF_bytes++;
-      else {
-        nr_of_FF_bytes = 0;
-        ff_flag = false;
-      }
-
-      if (nr_of_FF_bytes >= 3)
-        ff_flag = true;
-
-      response += (char) c;
-
-      if (recv_flag) {
-        if (response.indexOf(0x05) != -1) {
-          exit_flag = true;
-        }
-      }
-    }
-    if (exit_flag || ff_flag) {
-      break;
-    }
-  }
-
-  if (ff_flag)
-    response = response.substring(0, response.length() - 3);  // Remove last 3 0xFF
-
-  ret = response.length();
-  return ret;
+  return return_value;
 }
 
 void Nextion::soft_reset() { this->send_command_no_ack("rest"); }
-
-void Nextion::upload_tft() {
-  int old_baud = this->parent_->get_baud_rate();
-  if (this->has_updated_) {
-    ESP_LOGD(TAG, "Already Updated");
-    return;
-  }
-
-  if (this->is_updating_) {
-    ESP_LOGD(TAG, "Currently updating");
-    return;
-  }
-
-  if (!network_is_connected()) {
-    ESP_LOGD(TAG, "network is not connected");
-    return;
-  }
-
-  ESP_LOGD(TAG, "Updating tft from : %s", this->tft_url_.c_str());
-
-  this->is_updating_ = true;
-  this->total_ = 0;
-
-  HTTPClient http;
-  if (!http.begin(this->tft_url_.c_str())) {
-    this->is_updating_ = false;
-    ESP_LOGD(TAG, "connection failed");
-    return;
-  } else {
-    ESP_LOGD(TAG, "Connected");
-  }
-  http.addHeader("Range", "bytes=0-255");
-  const char *header_names[] = {"Content-Range"};
-  http.collectHeaders(header_names, 1);
-  ESP_LOGD(TAG, "Requesting URL: %s", this->tft_url_.c_str());
-
-  http.setReuse(true);
-  // try up to 5 times. DNS sometimes needs a second try or so
-  int tries = 0;
-  int code = http.GET();
-  while (code != 200 && code != 206 && tries < 5) {
-    for (int i = 0; i < 12; ++i)  // Need a decent delay and since we will be rebooting this shouldnt be an issue.
-      delay(40);
-    code = http.GET();
-    ++tries;
-  }
-  // Update the nextion display
-  if (code == 200 || code == 206) {
-    String content_range_string = http.header("Content-Range");
-    content_range_string.remove(0, 12);
-    int content_length = content_range_string.toInt();
-    http.end();  // End this HTTP call because we read all the data
-    delay(2);
-
-    ESP_LOGD(TAG, "Updating Nextion...");
-
-    String response = String("");
-    bool result;
-    char command[128];
-    sprintf(command, "whmi-wri %d,%d,0", content_length, this->parent_->get_baud_rate());
-    this->send_command_no_ack(command);
-    // Flush
-    this->flush();
-
-    this->recv_ret_string_(response, 800, true);  // normal response time is 400ms
-
-    // The Nextion display will, if it's ready to accept data, send a 0x05 byte.
-    if (response.indexOf(0x05) != -1) {
-      ESP_LOGD(TAG, "preparation for tft update done");
-    } else {
-      ESP_LOGD(TAG, "preparation for tft update failed");
-      this->sent_packets_ = 0;
-      this->is_updating_ = false;
-      return;
-    }
-    ESP_LOGD(TAG, "Start upload. File size is: %d bytes", content_length);
-    // Upload the received byte Stream to the nextion
-    result = this->upload_by_chunks_(content_length);
-
-    if (result) {
-      ESP_LOGD(TAG, "Succesfully updated Nextion! Sleep for 1600ms");
-    } else {
-      ESP_LOGD(TAG, "Error updating Nextion:");
-      this->sent_packets_ = 0;
-      this->is_updating_ = false;
-      return;
-    }
-
-    // end: wait(delay) for the nextion to finish the update process, send
-    // nextion reset command and end the serial connection to the nextion
-    // wait for the nextion to finish internal processes
-    delay(45);
-    delay(45);
-
-    // soft reset the nextion
-    this->soft_reset();
-
-    ESP_LOGD(TAG, "Nextion has been updated");
-    this->has_updated_ = true;
-    ESP.restart();
-  } else {
-    ESP_LOGD(TAG, "Nextion has NOT been updated, Bad HTTP status %d", code);
-  }
-  this->sent_packets_ = 0;
-  this->is_updating_ = false;
-
-}  // namespace nextion
 
 void Nextion::set_backlight_brightness(uint8_t brightness) { this->send_command_printf("dim=%u", brightness); }
 void Nextion::set_touch_sleep_timeout(uint16_t timeout) { this->send_command_printf("thsp=%u", timeout); }
