+++ esphome/components/nextion/nextion_sensor.h
@@ -11,14 +11,13 @@ class NextionSensor;
 class NextionSensor : public NextionComponent, public sensor::Sensor, public PollingComponent {
  public:
   NextionSensor(Nextion *nextion) { this->nextion_ = nextion; }
-
-  void process_sensor(uint8_t page_id, uint8_t component_id, float state) override;
-  void restore_from_nextion() override;
+  void process_sensor(char *variable_name, float state) override;
   void update() override;
+  void write_state(uint32_t state);
+  void nextion_setup() override;
 
  protected:
   Nextion *nextion_;
-  bool has_restored_ = false;
-};  // namespace nextion
+};
 }  // namespace nextion
 }  // namespace esphome
