+++ esphome/components/nextion/nextion_switch.h
@@ -2,23 +2,25 @@
 
 #include "esphome/core/component.h"
 #include "esphome/components/nextion/nextion.h"
-#include "esphome/components/sensor/sensor.h"
+#include "esphome/components/switch/switch.h"
 
 namespace esphome {
 namespace nextion {
 class NextionSwitch;
 
-class NextionSwitch : public NextionComponent, public switch_::Switch, public Component, public uart::UARTDevice {
+class NextionSwitch : public NextionComponent, public switch_::Switch, public PollingComponent {
  public:
   NextionSwitch(Nextion *nextion) { this->nextion_ = nextion; }
-  void process(uint8_t page_id, uint8_t component_id, bool on) override;
+  void process_bool(char *variable_name, bool on) override;
+  void update() override;
+  void nextion_setup() override;
 
  protected:
-  void send_command_no_ack(const char *command);
-  bool send_command_printf(const char *format, ...);
-
-  void write_state(bool state) override;
   Nextion *nextion_;
+  void write_state(bool state) override;
+
+  // void send_command_no_ack(const char *command);
+  // bool send_command_printf(const char *format, ...);
 };
 }  // namespace nextion
 }  // namespace esphome
