+++ esphome/components/nextion/nextion_textsensor.h
@@ -7,11 +7,14 @@
 namespace esphome {
 namespace nextion {
 class NextionTextSensor;
+
 class NextionTextSensor : public NextionComponent, public text_sensor::TextSensor, public PollingComponent {
  public:
   NextionTextSensor(Nextion *nextion) { this->nextion_ = nextion; }
-  void process_text(uint8_t page_id, uint8_t component_id, std::string state) override;
+  void process_text(char *variable_name, char *text_value) override;
   void update() override;
+  void write_state(std::string state);
+  void nextion_setup() override;
 
  protected:
   Nextion *nextion_;
